Refactor API/client structure
