+++ packages/app/src/api/synchronization/openapi/synchronizationAdjusted.json
